requirements.txt
firebase-admin==6.4.0
Flask==3.0.3
fpdf2==2.7.5
